double형으로 변경(키)

+++ DDOCDOC/src/ddocdoc/vo/ChildHeightVO.java
@@ -6,26 +6,26 @@ import java.sql.Date;
 public class ChildHeightVO implements Serializable{
 
 	private String he_num;
-	private int he_height;
+	private Double he_height;
 	private Date he_date;
 
 	private String ch_num;
 	
 	public ChildHeightVO() {}
 	
-	public ChildHeightVO(String he_num, int he_height) {
+	public ChildHeightVO(String he_num, Double he_height) {
 		super();
 		this.he_num = he_num;
 		this.he_height = he_height;
 	}
-	public ChildHeightVO(String he_num, int he_height, Date he_date) {
+	public ChildHeightVO(String he_num, Double he_height, Date he_date) {
 		super();
 		this.he_num = he_num;
 		this.he_height = he_height;
 		this.he_date = he_date;
 	}
 	
-	public ChildHeightVO(String he_num, int he_height, String ch_num, Date he_date) {
+	public ChildHeightVO(String he_num, Double he_height, String ch_num, Date he_date) {
 		super();
 		this.he_num = he_num;
 		this.he_height = he_height;
@@ -50,11 +50,11 @@ public class ChildHeightVO implements Serializable{
 		this.he_num = he_num;
 	}
 
-	public int getHe_height() {
+	public Double getHe_height() {
 		return he_height;
 	}
 
-	public void setHe_height(int he_height) {
+	public void setHe_height(Double he_height) {
 		this.he_height = he_height;
 	}
